feat: add applicant intake column and cleanup dashboard
