+++ app/components/Dashboard/index.ts
@@ -1,5 +1,6 @@
-export { default as DashboardTable } from './Table';
 export { default as Modal } from './Modal';
+export { default as Row } from './Row';
 export { default as StatusPill } from './StatusPill';
+export { default as Table } from './Table';
 export { default as Withdraw } from './Withdraw';
 export { default as X } from './XIcon';
